Prevent flickering logs by not using virtual repeat in static logs. Fixes #123
