+++ report-ng/app/src/components/abstract-log-view.ts
@@ -1,0 +1,52 @@
+/*
+ * Testerra
+ *
+ * (C) 2021, Mike Reiche,  T-Systems Multimedia Solutions GmbH, Deutsche Telekom AG
+ *
+ * Deutsche Telekom AG and all other contributors /
+ * copyright owners license this file to you under the Apache
+ * License, Version 2.0 (the "License"); you may not use this
+ * file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import {autoinject} from "aurelia-framework";
+import {StatusConverter} from "services/status-converter";
+import {ILogEntry} from "services/statistics-generator";
+
+@autoinject()
+export abstract class AbstractLogView {
+
+    constructor(
+        readonly statusConverter:StatusConverter,
+    ) {
+    }
+
+    protected toggleCause(logMessage:ILogEntry) {
+        if (logMessage.cause) {
+            logMessage.cause = null;
+        } else {
+            this.open(logMessage);
+        }
+    }
+
+    protected open(logMessage:ILogEntry) {
+        let msg = "";
+        logMessage.stackTrace.forEach(cause => {
+            if (msg.length > 0) {
+                msg += "<br>"
+            }
+            msg += (cause.message?.length>0?cause.message + "<br>":'') + cause.stackTraceElements.join("<br>");
+        });
+        logMessage.cause = msg;
+    }
+}
