+++ report-ng/app/src/components/log-view/log-line.html
@@ -1,0 +1,25 @@
+<!--
+  ~ Testerra
+  ~
+  ~ (C) 2021, Mike Reiche,  T-Systems Multimedia Solutions GmbH, Deutsche Telekom AG
+  ~
+  ~ Deutsche Telekom AG and all other contributors /
+  ~ copyright owners license this file to you under the Apache
+  ~ License, Version 2.0 (the "License"); you may not use this
+  ~ file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~ http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing,
+  ~ software distributed under the License is distributed on an
+  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+  ~ KIND, either express or implied.  See the License for the
+  ~ specific language governing permissions and limitations
+  ~ under the License.
+  -->
+<template>
+    <let logger.bind="statusConverter.separateNamespace(logMessage.loggerName)"></let>
+    <div class="indicator"><span if.bind="logMessage.stackTrace.length>0" click.delegate="toggleCause(logMessage)" class="clickable" innerhtml.bind="logMessage.cause?'-':'+'"></span></div>${logMessage.timestamp|dateFormat:"HH:mm:ss.SSS"}<span if.bind="logMessage.methodContext"> [<a route-href="route: method; params.bind: {methodId: logMessage.methodContext.contextValues.id}">${logMessage.threadName}</a>]</span> [${logMessage.type|logLevel}]: <span title="${logger.package}.${logger.class}">${logger.class}</span> - <span innerhtml.bind="logMessage.message|highlightText:searchRegexp"></span>
+    <div if.bind="logMessage.cause" innerhtml.bind="logMessage.cause|highlightText:searchRegexp"></div>
+</template>
