+++ report-ng/app/src/components/log-view/log-view.ts
@@ -19,101 +19,20 @@
  * under the License.
  */
 
+import {AbstractLogView} from "components/abstract-log-view";
 import {autoinject} from "aurelia-framework";
+import {StatusConverter} from "../../services/status-converter";
 import {bindable} from "aurelia-templating";
 import {bindingMode} from "aurelia-binding";
-import {StatusConverter} from "../../services/status-converter";
 import {ILogEntry} from "../../services/statistics-generator";
 
-@autoinject()
-export class LogView {
-
+export class LogView extends AbstractLogView {
     @bindable({defaultBindingMode: bindingMode.toView})
     class:string;
-
-    // @bindable({defaultBindingMode: bindingMode.fromView})
-    // logView:HTMLDivElement;
 
     @bindable({defaultBindingMode: bindingMode.toView})
     logMessages:ILogEntry[];
 
-    @bindable({defaultBindingMode: bindingMode.fromView})
-    readonly logView:LogView = this;
-
     @bindable({defaultBindingMode: bindingMode.toView})
     searchRegexp:RegExp;
-
-    private _logView:HTMLDivElement;
-    private _lastFoundIndex = -1;
-    private _initialScrollOffset:number;
-
-    constructor(
-        private _statusConverter:StatusConverter,
-    ) {
-    }
-
-    private _toggleCause(logMessage:ILogEntry) {
-        if (logMessage.cause) {
-            logMessage.cause = null;
-        } else {
-            this._open(logMessage);
-        }
-    }
-
-    private _open(logMessage:ILogEntry) {
-        let msg = "";
-        logMessage.stackTrace.forEach(cause => {
-            if (msg.length > 0) {
-                msg += "<br>"
-            }
-            msg += (cause.message?.length>0?cause.message + "<br>":'') + cause.stackTraceElements.join("<br>");
-        });
-        logMessage.cause = msg;
-    }
-
-    private _scrollToNextFound() {
-        if (this.searchRegexp && this._logView) {
-            const firstItem = this._logView.children.item(1) as HTMLDivElement;
-            const logMessagesToSearch = this.logMessages.slice(this._lastFoundIndex+1, -1);
-
-            const foundLogMessage = logMessagesToSearch.find(logMessage => {
-                const foundInMessage = logMessage.message.match(this.searchRegexp);
-                const foundInStackTrace = logMessage.stackTrace
-                    .flatMap(stackTrace => stackTrace.stackTraceElements)
-                    .filter(line => line.match(this.searchRegexp));
-
-                if (foundInStackTrace.length>0) {
-                    this._open(logMessage);
-                }
-
-                return foundInMessage || foundInStackTrace.length>0;
-            });
-
-            if (foundLogMessage) {
-                this._lastFoundIndex = this.logMessages.indexOf(foundLogMessage);
-                console.log("found at index", this._lastFoundIndex);
-                const scrollY = this._initialScrollOffset + (firstItem.clientHeight * this._lastFoundIndex);
-                console.log(this._initialScrollOffset);
-                console.log("scroll to ", this._initialScrollOffset, " + found index", this._lastFoundIndex, " * ", firstItem.clientHeight);
-                const options = {top: scrollY};
-
-                window.setTimeout(() => {
-                    window.scroll(options);
-                }, 1);
-            }
-        }
-    }
-
-    attached() {
-        this._initialScrollOffset = this._logView.getBoundingClientRect().top;
-    }
-
-    searchNext() {
-        this._scrollToNextFound();
-    }
-
-    resetSearch() {
-        console.log("reset search");
-        this._lastFoundIndex = -1;
-    }
 }
