+++ report-ng/app/src/components/logs/logs.html
@@ -49,12 +49,12 @@
             <mdc-layout-grid-cell span="4"></mdc-layout-grid-cell>
             <mdc-layout-grid-cell span="12" >
                 <mdc-linear-progress if.bind="_loading"></mdc-linear-progress>
-                <log-view else
+                <virtual-log-view else
                           class="logs"
                           log-view.bind="_logView"
                           log-messages.bind="_logMessages"
                           search-regexp.bind="_searchRegexp"
-                ></log-view>
+                ></virtual-log-view>
             </mdc-layout-grid-cell>
         </mdc-layout-grid-inner>
     </mdc-layout-grid>
